Fixed build issue of legacy NDIS-hooking driver

+++ include/Common.h
@@ -523,7 +523,7 @@ typedef struct _FAST_IO_SECTION
 typedef struct _INITIALIZE_FAST_IO_PARAMS
 {
 	PFAST_IO_SECTION header_ptr;
-	DWORD			 data_size;
+	ULONG			 data_size;
 }INITIALIZE_FAST_IO_PARAMS, *PINITIALIZE_FAST_IO_PARAMS;
 
 // --------------------------------------------------------------------------------
@@ -535,7 +535,7 @@ typedef struct _INITIALIZE_FAST_IO_PARAMS
 typedef struct _UNSORTED_READ_SEND_REQUEST
 {
 	PINTERMEDIATE_BUFFER*	packets;
-	DWORD					packets_num;
+	ULONG					packets_num;
 } UNSORTED_READ_SEND_REQUEST, *PUNSORTED_READ_SEND_REQUEST;
 
 // Restore Default Structure Packing
